Extract shared Python JITX parts package
