+++ jitx-py/alchitry-au1-level-shifter/src/components.py
@@ -16,163 +16,6 @@ def pad_attr(obj: object, name: str) -> Pad:
     return cast(Pad, getattr(obj, name))
 
 
-class SmallRectSmdPad(Pad):
-    shape = rectangle(0.3, 1.25)
-
-    def __init__(self):
-        self.soldermask = Soldermask(rectangle(0.4, 1.35))
-        self.paste = Paste(rectangle(0.4, 1.35))
-
-
-class LargeRectSmdPad(Pad):
-    shape = rectangle(2.0, 2.5)
-
-    def __init__(self):
-        self.soldermask = Soldermask(rectangle(2.1, 2.6))
-        self.paste = Paste(rectangle(2.1, 2.6))
-
-
-class ProbePthPad3mm(Pad):
-    shape = Circle(diameter=3.0)
-
-    def __init__(self):
-        self.cutout = Cutout(Circle(diameter=1.5))
-        self.soldermask = Soldermask(Circle(diameter=2.9))
-
-
-class GndTestpads(Component):
-    # Direct port of `components/GndTestpads.stanza`.
-    GND = Port()
-    reference_designator_prefix = "TP"
-    value = "~"
-
-    def __init__(self, *, diameter: float, width: float, height: float):
-        class _Landpattern(Landpattern):
-            p1 = ProbePthPad3mm().at(width / 2.0 - diameter, height / 2.0 - diameter)
-            p2 = ProbePthPad3mm().at(width / 2.0 - diameter, -height / 2.0 + diameter)
-            p3 = ProbePthPad3mm().at(-width / 2.0 + diameter, height / 2.0 - diameter)
-            p4 = ProbePthPad3mm().at(-width / 2.0 + diameter, -height / 2.0 + diameter)
-
-        self.landpattern = _Landpattern()
-        self.symbol = BoxSymbol(rows=[Row(left=PinGroup([self.GND]))], config=BoxConfig())
-        self.pad_mapping = PadMapping(
-            {
-                self.GND: [
-                    self.landpattern.p1,
-                    self.landpattern.p2,
-                    self.landpattern.p3,
-                    self.landpattern.p4,
-                ]
-            }
-        )
-
-
-class HDGC60PinFfcLandpattern(Landpattern):
-    # Footprint-level port of `components/_0_5K-1_2X-60PWB.stanza`.
-    def __init__(self):
-        self.signal_pads = [SmallRectSmdPad().at(14.75 - 0.5 * index, -2.096, rotate=180) for index in range(60)]
-        self.mount_pad_left = LargeRectSmdPad().at(-16.43, 0.527, rotate=180)
-        self.mount_pad_right = LargeRectSmdPad().at(16.43, 0.527, rotate=180)
-
-        self.ref_text = Silkscreen(Text(">REF", 0.5).at(-0.75, 4.477))
-        self.top_outline = Silkscreen(Polyline(0.254, [(-18.041, 2.644), (18.041, 2.644), (18.041, 1.247)]))
-        self.left_outline = Silkscreen(Polyline(0.254, [(-18.041, 2.644), (-18.041, 1.247), (-17.661, 1.247)]))
-        self.right_outline = Silkscreen(Polyline(0.254, [(17.661, 1.247), (18.041, 1.247)]))
-        self.lower_left = Silkscreen(Polyline(0.254, [(-17.406, -0.954), (-17.406, -1.547), (-15.131, -1.547)]))
-        self.lower_right = Silkscreen(Polyline(0.254, [(15.131, -1.547), (17.406, -1.547), (17.406, -0.954)]))
-        self.courtyard = Courtyard(rectangle(36.336, 5.543))
-
-
-class HDGC60PinFfc(Component):
-    # Raw connector component port of `components/_0_5K-1_2X-60PWB.stanza`.
-    p = [Port() for _ in range(62)]
-
-    manufacturer = "HDGC"
-    mpn = "0.5K-1.2X-60PWB"
-    datasheet = "https://wmsc.lcsc.com/wmsc/upload/file/pdf/v2/lcsc/2111151930_HDGC-0-5K-1-2X-60PWB_C2919552.pdf"
-    description = "60-pin 0.5 mm bottom-contact FFC/FPC connector"
-    reference_designator_prefix = "J"
-    value = "0.5K-1.2X-60PWB"
-
-    def __init__(self):
-        self.landpattern = HDGC60PinFfcLandpattern()
-        rows = [Row(left=PinGroup([port])) for port in self.p[:60]]
-        rows.append(Row(left=PinGroup([self.p[60]]), right=PinGroup([self.p[61]])))
-        self.symbol = BoxSymbol(rows=rows, config=BoxConfig(group_spacing=1))
-
-        mapping = {self.p[i]: self.landpattern.signal_pads[i] for i in range(60)}
-        mapping[self.p[60]] = self.landpattern.mount_pad_left
-        mapping[self.p[61]] = self.landpattern.mount_pad_right
-        self.pad_mapping = PadMapping(mapping)
-
-
-class SaleaeProbePthPad(Pad):
-    shape = Circle(diameter=1.4)
-
-    def __init__(self):
-        self.cutout = Cutout(Circle(diameter=1.0))
-        self.soldermask = Soldermask(Circle(diameter=1.5))
-
-
-class SaleaeProbeHeaderLandpattern(Landpattern):
-    # Exact port of the archived `PIN_HDR_8` footprint reused by Saleae boards.
-    def __init__(self):
-        self.p1 = SaleaeProbePthPad().at(-1.27, 3.81)
-        self.p2 = SaleaeProbePthPad().at(1.27, 3.81)
-        self.p3 = SaleaeProbePthPad().at(-1.27, 1.27)
-        self.p4 = SaleaeProbePthPad().at(1.27, 1.27)
-        self.p5 = SaleaeProbePthPad().at(-1.27, -1.27)
-        self.p6 = SaleaeProbePthPad().at(1.27, -1.27)
-        self.p7 = SaleaeProbePthPad().at(-1.27, -3.81)
-        self.p8 = SaleaeProbePthPad().at(1.27, -3.81)
-        self.outline = Silkscreen(
-            Polyline(
-                0.153,
-                [(-2.54, -5.08), (-2.54, 5.08), (2.54, 5.08), (2.54, -5.08), (-2.54, -5.08)],
-            )
-        )
-
-
-class SaleaeProbeHeader2x4(Component):
-    GND = Port()
-    p0 = Port()
-    p1 = Port()
-    p2 = Port()
-    p3 = Port()
-
-    reference_designator_prefix = "J"
-    manufacturer = "Generic"
-    mpn = "generic-2x4-2.54mm-th"
-    description = "Generic 2x4 2.54 mm Saleae-compatible probe header"
-    value = "8X2-pin-header"
-
-    def __init__(self):
-        self.landpattern = SaleaeProbeHeaderLandpattern()
-        self.symbol = BoxSymbol(
-            rows=[
-                Row(left=PinGroup([self.GND]), right=PinGroup([self.p3])),
-                Row(right=PinGroup([self.p2])),
-                Row(right=PinGroup([self.p1])),
-                Row(right=PinGroup([self.p0])),
-            ],
-            config=BoxConfig(group_spacing=2),
-        )
-        self.pad_mapping = PadMapping(
-            {
-                self.GND: (
-                    self.landpattern.p1,
-                    self.landpattern.p3,
-                    self.landpattern.p5,
-                    self.landpattern.p7,
-                ),
-                self.p0: self.landpattern.p8,
-                self.p1: self.landpattern.p6,
-                self.p2: self.landpattern.p4,
-                self.p3: self.landpattern.p2,
-            }
-        )
-
-
 class HeaderPthPad(Pad):
     shape = Circle(diameter=1.4)
