+++ jitx-py/retrobus-jitx-parts/.gitignore
@@ -1,0 +1,4 @@
+.venv/
+__pycache__/
+*.pyc
+.ruff_cache/
