+++ jitx-py/retrobus-jitx-parts/README.md
@@ -1,0 +1,3 @@
+# retrobus-jitx-parts
+
+Shared reusable JITX Python parts for the `jitx-py` board projects.
