+++ jitx-py/retrobus-jitx-parts/pyproject.toml
@@ -3,9 +3,9 @@ requires = ["hatchling>=1.27.0,<2.0"]
 build-backend = "hatchling.build"
 
 [project]
-name = "saleae-dslab-adapter-py"
+name = "retrobus-jitx-parts"
 version = "0.1.0"
-description = "Python JITX port scaffold for the retrobus saleae-dslab-adapter board"
+description = "Shared reusable JITX Python parts for retrobus jitx-py projects"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
@@ -13,20 +13,16 @@ dependencies = [
   "jitxlib-parts>=1.0.0,<2",
   "jitxlib-standard>=4.0.0,<5",
   "jitxexamples-components>=1.0.0,<2",
-  "retrobus-jitx-parts",
 ]
 
 [tool.hatch.build.targets.wheel]
-packages = ["src"]
+packages = ["src/retrobus_jitx_parts"]
 
 [project.optional-dependencies]
 dev = [
   "ruff",
   "ty",
 ]
-
-[tool.uv.sources]
-retrobus-jitx-parts = { path = "../retrobus-jitx-parts", editable = true }
 
 [tool.ruff]
 target-version = "py312"
