+++ jitx-py/sharp-pc-e500-ram-card/pyproject.toml
@@ -13,6 +13,7 @@ dependencies = [
   "jitxlib-parts>=1.0.0,<2",
   "jitxlib-standard>=4.0.0,<5",
   "jitxexamples-components>=1.0.0,<2",
+  "retrobus-jitx-parts",
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -23,6 +24,9 @@ dev = [
   "ruff",
   "ty",
 ]
+
+[tool.uv.sources]
+retrobus-jitx-parts = { path = "../retrobus-jitx-parts", editable = true }
 
 [tool.ruff]
 target-version = "py312"
