+++ jitx-py/sharp-pc-e500-ram-card/src/components.py
@@ -1,67 +1,11 @@
 from __future__ import annotations
 
 from jitx.component import Component
-from jitx.feature import Courtyard, Paste, Silkscreen, Soldermask
+from jitx.feature import Paste, Soldermask
 from jitx.landpattern import Landpattern, Pad, PadMapping
 from jitx.net import Port
 from jitx.shapes.composites import rectangle
-from jitx.shapes.primitive import Polyline, Text
 from jitxlib.symbols.box import BoxConfig, BoxSymbol, PinGroup, Row
-
-
-class SmallRectSmdPad(Pad):
-    shape = rectangle(0.3, 1.25)
-
-    def __init__(self):
-        self.soldermask = Soldermask(rectangle(0.4, 1.35))
-        self.paste = Paste(rectangle(0.4, 1.35))
-
-
-class LargeRectSmdPad(Pad):
-    shape = rectangle(2.0, 2.5)
-
-    def __init__(self):
-        self.soldermask = Soldermask(rectangle(2.1, 2.6))
-        self.paste = Paste(rectangle(2.1, 2.6))
-
-
-class HDGC60PinFfcLandpattern(Landpattern):
-    # Footprint-level port of `components/_0_5K-1_2X-60PWB.stanza`.
-    def __init__(self):
-        self.signal_pads = [SmallRectSmdPad().at(14.75 - 0.5 * index, -2.096, rotate=180) for index in range(60)]
-        self.mount_pad_left = LargeRectSmdPad().at(-16.43, 0.527, rotate=180)
-        self.mount_pad_right = LargeRectSmdPad().at(16.43, 0.527, rotate=180)
-
-        self.ref_text = Silkscreen(Text(">REF", 0.5).at(-0.75, 4.477))
-        self.top_outline = Silkscreen(Polyline(0.254, [(-18.041, 2.644), (18.041, 2.644), (18.041, 1.247)]))
-        self.left_outline = Silkscreen(Polyline(0.254, [(-18.041, 2.644), (-18.041, 1.247), (-17.661, 1.247)]))
-        self.right_outline = Silkscreen(Polyline(0.254, [(17.661, 1.247), (18.041, 1.247)]))
-        self.lower_left = Silkscreen(Polyline(0.254, [(-17.406, -0.954), (-17.406, -1.547), (-15.131, -1.547)]))
-        self.lower_right = Silkscreen(Polyline(0.254, [(15.131, -1.547), (17.406, -1.547), (17.406, -0.954)]))
-        self.courtyard = Courtyard(rectangle(36.336, 5.543))
-
-
-class HDGC60PinFfc(Component):
-    # Raw connector component port of `components/_0_5K-1_2X-60PWB.stanza`.
-    p = [Port() for _ in range(62)]
-
-    manufacturer = "HDGC"
-    mpn = "0.5K-1.2X-60PWB"
-    datasheet = "https://wmsc.lcsc.com/wmsc/upload/file/pdf/v2/lcsc/2111151930_HDGC-0-5K-1-2X-60PWB_C2919552.pdf"
-    description = "60-pin 0.5 mm bottom-contact FFC/FPC connector"
-    reference_designator_prefix = "J"
-    value = "0.5K-1.2X-60PWB"
-
-    def __init__(self):
-        self.landpattern = HDGC60PinFfcLandpattern()
-        rows = [Row(left=PinGroup([port])) for port in self.p[:60]]
-        rows.append(Row(left=PinGroup([self.p[60]]), right=PinGroup([self.p[61]])))
-        self.symbol = BoxSymbol(rows=rows, config=BoxConfig(group_spacing=1))
-
-        mapping = {self.p[i]: self.landpattern.signal_pads[i] for i in range(60)}
-        mapping[self.p[60]] = self.landpattern.mount_pad_left
-        mapping[self.p[61]] = self.landpattern.mount_pad_right
-        self.pad_mapping = PadMapping(mapping)
 
 
 class BigRamCardPad(Pad):
